+++ labs-ios-starter/Surveys/View Controllers/TopicQuestionsViewController.swift
@@ -94,8 +94,8 @@ class TopicQuestionsViewController: UIViewController {
         }
 
         let selected = contextSegmentControl.selectedSegmentIndex
-
-        topicController.postTopic(with: topicName, contextId: Int64(selected), questions: topicController.questions) { result in
+        #warning("Error 500 here")
+        topicController.postTopic(with: topicName, contextId: selected, questions: topicController.questions) { result in
             switch result {
             case .success(let joinCode):
                 self.presentSimpleAlert(with: "Topic Posted!", message: "Your join code is \(joinCode)", preferredStyle: .alert, dismissText: "Got it!")
